# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LEXANDR — Deep Research Intelligence API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AnthropicConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface AnthropicMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateAnthropicConversationBody {
  title: string;
}

export interface SendAnthropicMessageBody {
  content: string;
}

export interface AnthropicConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: AnthropicMessage[];
}

export interface AnthropicError {
  error: string;
}

export interface ResearchSession {
  id: number;
  topic: string;
  domain: string;
  /** @nullable */
  description: string | null;
  messageCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface ResearchMessage {
  id: number;
  sessionId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface ResearchSessionWithMessages {
  id: number;
  topic: string;
  domain: string;
  /** @nullable */
  description: string | null;
  messageCount: number;
  createdAt: string;
  updatedAt: string;
  messages: ResearchMessage[];
}

export interface CreateResearchSessionBody {
  topic: string;
  domain: string;
  description?: string;
}

export interface SendResearchMessageBody {
  content: string;
}

export interface ResearchReport {
  id: number;
  sessionId: number;
  topic: string;
  domain: string;
  summary: string;
  scientificBasis: string;
  marketAnalysis: string;
  technicalRecommendations: string;
  strategicInsights: string;
  riskAnalysis: string;
  frameworks: string;
  sources: string;
  createdAt: string;
}

export interface DomainCount {
  domain: string;
  count: number;
}

export interface ResearchStats {
  totalSessions: number;
  totalMessages: number;
  totalReports: number;
  sessionsByDomain: DomainCount[];
  recentSessions: ResearchSession[];
}

export interface ResearchError {
  error: string;
}