# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LEXANDR — Deep Research Intelligence API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: anthropic
    description: Anthropic AI chat operations
  - name: research
    description: LEXANDR research session operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /anthropic/conversations:
    get:
      operationId: listAnthropicConversations
      tags: [anthropic]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnthropicConversation"
    post:
      operationId: createAnthropicConversation
      tags: [anthropic]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnthropicConversationBody"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnthropicConversation"
  /anthropic/conversations/{id}:
    get:
      operationId: getAnthropicConversation
      tags: [anthropic]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnthropicConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnthropicError"
    delete:
      operationId: deleteAnthropicConversation
      tags: [anthropic]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnthropicError"
  /anthropic/conversations/{id}/messages:
    get:
      operationId: listAnthropicMessages
      tags: [anthropic]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnthropicMessage"
    post:
      operationId: sendAnthropicMessage
      tags: [anthropic]
      summary: Send a message and receive an AI response (SSE stream)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendAnthropicMessageBody"
      responses:
        "200":
          description: SSE stream of assistant response chunks
          content:
            text/event-stream: {}

  /research/sessions:
    get:
      operationId: listResearchSessions
      tags: [research]
      summary: List all LEXANDR research sessions
      responses:
        "200":
          description: List of research sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResearchSession"
    post:
      operationId: createResearchSession
      tags: [research]
      summary: Create a new research session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResearchSessionBody"
      responses:
        "201":
          description: Created research session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchSession"

  /research/sessions/{id}:
    get:
      operationId: getResearchSession
      tags: [research]
      summary: Get research session with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Research session with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchSessionWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchError"
    delete:
      operationId: deleteResearchSession
      tags: [research]
      summary: Delete a research session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchError"

  /research/sessions/{id}/messages:
    post:
      operationId: sendResearchMessage
      tags: [research]
      summary: Send a research query and receive LEXANDR response (SSE stream)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendResearchMessageBody"
      responses:
        "200":
          description: SSE stream of LEXANDR research response
          content:
            text/event-stream: {}

  /research/sessions/{id}/report:
    post:
      operationId: generateResearchReport
      tags: [research]
      summary: Generate a structured research report for a session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Generated research report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchReport"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchError"

  /research/reports:
    get:
      operationId: listResearchReports
      tags: [research]
      summary: List all generated research reports
      responses:
        "200":
          description: List of research reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResearchReport"

  /research/reports/{id}:
    get:
      operationId: getResearchReport
      tags: [research]
      summary: Get a specific research report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Research report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchReport"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchError"

  /research/stats:
    get:
      operationId: getResearchStats
      tags: [research]
      summary: Get aggregate research statistics
      responses:
        "200":
          description: Research statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    AnthropicConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt
    AnthropicMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    CreateAnthropicConversationBody:
      type: object
      properties:
        title:
          type: string
      required:
        - title
    SendAnthropicMessageBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content
    AnthropicConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/AnthropicMessage"
      required:
        - id
        - title
        - createdAt
        - messages
    AnthropicError:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    ResearchSession:
      type: object
      properties:
        id:
          type: integer
        topic:
          type: string
        domain:
          type: string
        description:
          type: ["string", "null"]
        messageCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - topic
        - domain
        - description
        - messageCount
        - createdAt
        - updatedAt

    ResearchSessionWithMessages:
      type: object
      properties:
        id:
          type: integer
        topic:
          type: string
        domain:
          type: string
        description:
          type: ["string", "null"]
        messageCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ResearchMessage"
      required:
        - id
        - topic
        - domain
        - description
        - messageCount
        - createdAt
        - updatedAt
        - messages

    ResearchMessage:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sessionId
        - role
        - content
        - createdAt

    CreateResearchSessionBody:
      type: object
      properties:
        topic:
          type: string
        domain:
          type: string
        description:
          type: string
      required:
        - topic
        - domain

    SendResearchMessageBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content

    ResearchReport:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        topic:
          type: string
        domain:
          type: string
        summary:
          type: string
        scientificBasis:
          type: string
        marketAnalysis:
          type: string
        technicalRecommendations:
          type: string
        strategicInsights:
          type: string
        riskAnalysis:
          type: string
        frameworks:
          type: string
        sources:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sessionId
        - topic
        - domain
        - summary
        - scientificBasis
        - marketAnalysis
        - technicalRecommendations
        - strategicInsights
        - riskAnalysis
        - frameworks
        - sources
        - createdAt

    ResearchStats:
      type: object
      properties:
        totalSessions:
          type: integer
        totalMessages:
          type: integer
        totalReports:
          type: integer
        sessionsByDomain:
          type: array
          items:
            $ref: "#/components/schemas/DomainCount"
        recentSessions:
          type: array
          items:
            $ref: "#/components/schemas/ResearchSession"
      required:
        - totalSessions
        - totalMessages
        - totalReports
        - sessionsByDomain
        - recentSessions

    DomainCount:
      type: object
      properties:
        domain:
          type: string
        count:
          type: integer
      required:
        - domain
        - count

    ResearchError:
      type: object
      properties:
        error:
          type: string
      required:
        - error