# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LEXANDR — Deep Research Intelligence API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all conversations
 */
export const ListAnthropicConversationsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListAnthropicConversationsResponse = zod.array(
  ListAnthropicConversationsResponseItem,
);

/**
 * @summary Create a new conversation
 */
export const CreateAnthropicConversationBody = zod.object({
  title: zod.string(),
});

/**
 * @summary Get conversation with messages
 */
export const GetAnthropicConversationParams = zod.object({
  id: zod.coerce.number(),
});

export const GetAnthropicConversationResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
  messages: zod.array(
    zod.object({
      id: zod.number(),
      conversationId: zod.number(),
      role: zod.string(),
      content: zod.string(),
      createdAt: zod.coerce.date(),
    }),
  ),
});

/**
 * @summary Delete a conversation
 */
export const DeleteAnthropicConversationParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List messages in a conversation
 */
export const ListAnthropicMessagesParams = zod.object({
  id: zod.coerce.number(),
});

export const ListAnthropicMessagesResponseItem = zod.object({
  id: zod.number(),
  conversationId: zod.number(),
  role: zod.string(),
  content: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListAnthropicMessagesResponse = zod.array(
  ListAnthropicMessagesResponseItem,
);

/**
 * @summary Send a message and receive an AI response (SSE stream)
 */
export const SendAnthropicMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const SendAnthropicMessageBody = zod.object({
  content: zod.string(),
});

/**
 * @summary List all LEXANDR research sessions
 */
export const ListResearchSessionsResponseItem = zod.object({
  id: zod.number(),
  topic: zod.string(),
  domain: zod.string(),
  description: zod.string().nullable(),
  messageCount: zod.number(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListResearchSessionsResponse = zod.array(
  ListResearchSessionsResponseItem,
);

/**
 * @summary Create a new research session
 */
export const CreateResearchSessionBody = zod.object({
  topic: zod.string(),
  domain: zod.string(),
  description: zod.string().optional(),
});

/**
 * @summary Get research session with messages
 */
export const GetResearchSessionParams = zod.object({
  id: zod.coerce.number(),
});

export const GetResearchSessionResponse = zod.object({
  id: zod.number(),
  topic: zod.string(),
  domain: zod.string(),
  description: zod.string().nullable(),
  messageCount: zod.number(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
  messages: zod.array(
    zod.object({
      id: zod.number(),
      sessionId: zod.number(),
      role: zod.string(),
      content: zod.string(),
      createdAt: zod.coerce.date(),
    }),
  ),
});

/**
 * @summary Delete a research session
 */
export const DeleteResearchSessionParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Send a research query and receive LEXANDR response (SSE stream)
 */
export const SendResearchMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const SendResearchMessageBody = zod.object({
  content: zod.string(),
});

/**
 * @summary Generate a structured research report for a session
 */
export const GenerateResearchReportParams = zod.object({
  id: zod.coerce.number(),
});

export const GenerateResearchReportResponse = zod.object({
  id: zod.number(),
  sessionId: zod.number(),
  topic: zod.string(),
  domain: zod.string(),
  summary: zod.string(),
  scientificBasis: zod.string(),
  marketAnalysis: zod.string(),
  technicalRecommendations: zod.string(),
  strategicInsights: zod.string(),
  riskAnalysis: zod.string(),
  frameworks: zod.string(),
  sources: zod.string(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary List all generated research reports
 */
export const ListResearchReportsResponseItem = zod.object({
  id: zod.number(),
  sessionId: zod.number(),
  topic: zod.string(),
  domain: zod.string(),
  summary: zod.string(),
  scientificBasis: zod.string(),
  marketAnalysis: zod.string(),
  technicalRecommendations: zod.string(),
  strategicInsights: zod.string(),
  riskAnalysis: zod.string(),
  frameworks: zod.string(),
  sources: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListResearchReportsResponse = zod.array(
  ListResearchReportsResponseItem,
);

/**
 * @summary Get a specific research report
 */
export const GetResearchReportParams = zod.object({
  id: zod.coerce.number(),
});

export const GetResearchReportResponse = zod.object({
  id: zod.number(),
  sessionId: zod.number(),
  topic: zod.string(),
  domain: zod.string(),
  summary: zod.string(),
  scientificBasis: zod.string(),
  marketAnalysis: zod.string(),
  technicalRecommendations: zod.string(),
  strategicInsights: zod.string(),
  riskAnalysis: zod.string(),
  frameworks: zod.string(),
  sources: zod.string(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Get aggregate research statistics
 */
export const GetResearchStatsResponse = zod.object({
  totalSessions: zod.number(),
  totalMessages: zod.number(),
  totalReports: zod.number(),
  sessionsByDomain: zod.array(
    zod.object({
      domain: zod.string(),
      count: zod.number(),
    }),
  ),
  recentSessions: zod.array(
    zod.object({
      id: zod.number(),
      topic: zod.string(),
      domain: zod.string(),
      description: zod.string().nullable(),
      messageCount: zod.number(),
      createdAt: zod.coerce.date(),
      updatedAt: zod.coerce.date(),
    }),
  ),
});