# lib/api-zod/src/generated/types/createResearchSessionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LEXANDR — Deep Research Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateResearchSessionBody {
  topic: string;
  domain: string;
  description?: string;
}