# lib/api-zod/src/generated/types/researchReport.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LEXANDR — Deep Research Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface ResearchReport {
  id: number;
  sessionId: number;
  topic: string;
  domain: string;
  summary: string;
  scientificBasis: string;
  marketAnalysis: string;
  technicalRecommendations: string;
  strategicInsights: string;
  riskAnalysis: string;
  frameworks: string;
  sources: string;
  createdAt: Date;
}