# lib/api-zod/src/generated/types/researchSession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LEXANDR — Deep Research Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface ResearchSession {
  id: number;
  topic: string;
  domain: string;
  /** @nullable */
  description: string | null;
  messageCount: number;
  createdAt: Date;
  updatedAt: Date;
}